+++ packages/terra-list/tests/nightwatch/single-select-list/SingleSelectListNoItems.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import SingleSelectList from '../../../lib/SingleSelectList';
-
-const list = () => (
-  <SingleSelectList />
- );
-
-export default list;